requirements.txt
asgiref==3.5.1
beautifulsoup4==4.11.1
bootstrap4==0.1.0
Django==4.0.4
django-bootstrap4==22.1
django-ckeditor==6.4.0
django-js-asset==2.0.0
Pillow==9.1.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
